[nan-1762] remove casting
