+++ flows.yaml
@@ -1601,6 +1601,8 @@ integrations:
                     - https://www.googleapis.com/auth/admin.directory.user.readonly
             workspace-users:
                 runs: every hour
+                input: Metadata
+                auto_start: false
                 output: User
                 sync_type: full
                 endpoint: GET /google/workspace-users
@@ -1613,6 +1615,11 @@ integrations:
                     - https://www.googleapis.com/auth/admin.directory.user.readonly
                     - https://www.googleapis.com/auth/admin.directory.user.security
         models:
+            Metadata:
+                orgsToSync: OrgToSync[]
+            OrgToSync:
+                id: string
+                path: string
             OrganizationalUnit:
                 id: string
                 name: string
