+++ integrations/google/nango.yaml
@@ -12,6 +12,8 @@ integrations:
                     - https://www.googleapis.com/auth/admin.directory.user.readonly
             workspace-users:
                 runs: every hour
+                input: Metadata
+                auto_start: false
                 output: User
                 sync_type: full
                 endpoint: GET /google/workspace-users
@@ -24,6 +26,11 @@ integrations:
                     - https://www.googleapis.com/auth/admin.directory.user.readonly
                     - https://www.googleapis.com/auth/admin.directory.user.security
 models:
+    Metadata:
+        orgsToSync: OrgToSync[]
+    OrgToSync:
+        id: string
+        path: string
     OrganizationalUnit:
         id: string
         name: string
